Milan wrote:
I've noticed a few things that could easily be improved on GNUnet's
webpage (see attached patch):
- in "GNU's", the apostrophe is in the wrong direction
- the list of supported OSes is too long :-) so it could be nice to
insert a return before it, so that the left frame is less wide
- the 'Contact' section at the bottom of the homepage is useless since
it does not say much, and a page exists for that
- gnunet-qt should be mentioned with gnunet-gtk

I'm not used to the site's framework, esp. concerning a new <br> line
into a PHP/i18nHTML system that I cannot test, so I prefer you to review
this before.

Another idea is that the 'About GNUnet' section could be improved: it
only explains the economical model of GNUnet, not how it (roughly)
achieves anonynimity. This is quickly explained below, in 'Why GNUnet?',
but should be developed sooner. I've seen many times sites merely
copying this description when reviewing anonymous P2P networks, and this
explains very poorly GNUnet. Would you find two sentences for that?

Also, I've thought we could create a new section 'Getting started' (or
the like) aiming at people that only want to get GNUnet working. The FAQ
is nice, but the Documentation is aiming at advanced users that want to
understand the framework or to build from source, etc. For that,
'Papers' could be moved to the Documentation page, and replaced in the
header by this new section. I can do this but to write it first a book
in Drupal would be nice.


Congratulations for 0.8.0!