Changes mp3 extractor to discard frames with frame_size <= 0. Obviously, this is quite hacky, but:   A) Files with 0 bitrate are rare.   B) At least it won't trigger as easily on random binary data
Also, makes mp3 extractor check for at least 2 valid frames in a row.
That is, if frame size is not 0, then the next frame header after the
first one we found must also be valid, otherwise this is not
considered to be an mp3 file. This is the best i can do with such a
short and simple patch, further improvements are only possible with
something more complex, i think.