LRN: 1) Use PATH_SEPARATOR_STR instead of ":" to tokenize prefix
2) Correctly close cpipe_out and cpipe_in on W32

3) Use page-backed memory mapping instead of a file-backed one (give
the child process a name of the mapping instead of a file name).

4) Add RundllEntryPointA() (which calls RundllEntryPoint() anyway),
doesn't work on NT 6.1 without it.

5) Pass duplicated HANDLEs to the child, not CRT file IDs (apparently,
CRT file IDs are not inherited).

6) Call CreateProcess correctly (so that it finds rundll32 in PATH).

7) Write data into the child's input pipe AFTER the child starts
(otherwise there is a probability of making a blocking write and
hanging up).

AFAICS, passing the test requires LIBEXTRACTOR_PREFIX to be set and
pointing to $GNUNET_PREFIX/lib/libextractor directory